Use space for delegations
