+++ src/strategies/erc20-balance-of-delegation/index.ts
@@ -1,5 +1,5 @@
 import { strategy as erc20BalanceOfStrategy } from '../erc20-balance-of';
-import { subgraphRequest, SNAPSHOT_SUBGRAPH_URL } from '../../utils';
+import { getDelegations } from '../../plugins/delegation/utils';
 
 export const author = 'bonustrack';
 export const version = '0.1.0';
@@ -12,54 +12,33 @@ export async function strategy(
   options,
   snapshot
 ) {
-  const params = {
-    delegations: {
-      __args: {
-        where: {
-          delegate_in: addresses.map((address) => address.toLowerCase()),
-          space: ''
-        },
-        first: 1000
-      },
-      delegator: true,
-      space: true,
-      delegate: true
-    }
-  };
-  if (snapshot !== 'latest') {
-    // @ts-ignore
-    params.delegations.__args.block = { number: snapshot };
-  }
-  const result = await subgraphRequest(SNAPSHOT_SUBGRAPH_URL[network], params);
-  if (result && result.delegations) {
-    const delegators = result.delegations.map(
-      (delegation) => delegation.delegator
-    );
-    const delegatorsByAddress = {};
-    result.delegations.forEach((delegation) => {
-      if (!delegatorsByAddress[delegation.delegate])
-        delegatorsByAddress[delegation.delegate] = [];
-      delegatorsByAddress[delegation.delegate].push(delegation.delegator);
-    });
-    const score = await erc20BalanceOfStrategy(
-      space,
-      network,
-      provider,
-      delegators,
-      options,
-      snapshot
-    );
-    return Object.fromEntries(
-      addresses.map((address) => {
-        const addressScore = delegatorsByAddress[address.toLowerCase()]
-          ? delegatorsByAddress[address.toLowerCase()].reduce(
-              (a, b) => a + score[b],
-              0
-            )
-          : 0;
-        return [address, addressScore];
-      })
-    );
-  }
-  return {};
+  const delegations = await getDelegations(
+    space,
+    network,
+    provider,
+    addresses,
+    options,
+    snapshot
+  );
+  if (Object.keys(delegations).length === 0) return {};
+
+  const score = await erc20BalanceOfStrategy(
+    space,
+    network,
+    provider,
+    Object.values(delegations).reduce((a: string[], b: string[]) =>
+      a.concat(b)
+    ),
+    options,
+    snapshot
+  );
+
+  return Object.fromEntries(
+    addresses.map((address) => {
+      const addressScore = delegations[address.toLowerCase()]
+        ? delegations[address.toLowerCase()].reduce((a, b) => a + score[b], 0)
+        : 0;
+      return [address, addressScore];
+    })
+  );
 }
